python/pyproject.toml
[project]
name = "rekolto-python"
version = "0.1.0"
description = "Python services for Rekolto (PageIndex)"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "google-generativeai>=0.8.6",
]
